+++ global-perspectives-starter/frontend/src/App.jsx
@@ -21,6 +21,7 @@ import AuthCallback from './components/AuthCallback';
 import Account from './components/Account';
 import WhitepaperPage from './components/WhitepaperPage';
 import WorldMapV2 from './components/WorldMapV2';
+import SituationHome from './components/SituationHome';
 import EconomyPage from './components/EconomyPage';
 import AnalysisStudio from './components/AnalysisStudio';
 import MembershipPage from './components/MembershipPage';
@@ -98,7 +99,8 @@ export default function App() {
             <ErrorBoundary>
             <Routes>
               <Route path="/" element={<Home />} />
-              <Route path="/map" element={<WorldMapV2 />} />
+              <Route path="/map" element={<SituationHome />} />
+              <Route path="/map-legacy" element={<WorldMapV2 />} />
               <Route path="/privacy" element={<PrivacyTerms />} />
               <Route path="/about" element={<AboutContact />} />
               <Route path="/disclosures" element={<Disclosures />} />